requirements.txt
asgiref==3.2.10
Django==3.1.1
django-admin-autocomplete-filter==0.6.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.11.1
mysql==0.0.2
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
reportlab==3.5.51
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.10
